requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.0.0

# Computer Vision
opencv-python>=4.8.0
ultralytics>=8.0.0  # YOLOv8

# Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Interface
gradio>=4.0.0

# Configuration
pyyaml>=6.0

# Dataset Download (optional)
kaggle>=1.5.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Development (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
